requirements.txt
selenium
pymongo
python-dotenv
Flask-PyMongo